#68 improved winning move algorithm
improved the efficiency and completeness for edge cases in the winning move algorithm (winCheck).  Also, since the game is no longer in the terminal I changed the range used for column in main.py which just made things look better.  The change in the column range was also reflected in a change to the list used to check for a draw.  The game should now be fully functional locally without any bugs.